requirements.txt
openai
python-telegram-bot
pytube
pydub
ffmpeg
yt-dlp
dotenv